requirements.txt

fastapi==0.124.0
uvicorn[standard]==0.38.0
starlette==0.50.0
pydantic==2.12.4
python-multipart>=0.0.9


pymongo==4.5.0
motor==3.3.2
python-dotenv==1.0.0


ultralytics==8.3.0
opencv-python==4.10.0.84
matplotlib==3.8.0
numpy==1.26.0
pillow==11.1.0

transformers==4.40.0
torch==2.3.1
torchvision==0.18.1
sentencepiece==0.2.0
accelerate==0.30.1
google-generativeai==0.3.0
openai==1.30.1


streamlit==1.43.1
plotly==5.23.0
pandas==2.2.3


requests==2.32.3
httpx==0.27.0  

geopy==2.4.1
osmnx==1.7.0
networkx==3.2.1
folium==0.14.0
shapely==2.0.2

supabase==2.25.1

celery==5.3.4
redis==5.0.1
flower==2.0.1
slowapi>=0.1.9


streamlit-option-menu==0.3.6
